feat: update error messages to Portuguese and refactor routes

+++ src/application/use-cases/auth/DeleteUserUseCase.ts
@@ -4,6 +4,9 @@ export class DeleteUserUseCase {
     constructor(private userRepository: IUserRepository) {}
 
     async execute(id: string): Promise<boolean> {
+        if (this.userRepository.findById(id) === null) {
+            throw new Error('Usuário não encontrado ou inexistente');
+        }
         return await this.userRepository.delete(id);
     }
 }
